[AutoWS] Clean up 2-CTA dependency test structure
Move the core Triton tl.dot 2-CTA runtime coverage out of the TLX dot test, tighten the per-MMA sync lit checks, and factor matmul dependency tracing into reusable helpers.

Authored with Claude.
